refactor: Follow Api 주석 추가
