+++ Moyoy-Infra/src/main/java/com/moyoy/infra/external/github/user/GithubUserFeignClient.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-import com.moyoy.infra.external.github.support.GithubFeignConfig;
+import com.moyoy.infra.external.github.support.config.GithubFeignConfig;
 
 import feign.Response;
